docs_generator/requirements.txt
# File: requirements.txt
networkx==2.8.4
matplotlib==3.5.2
astroid==2.15.5
Jinja2==3.1.2
python-docx==0.8.11

